requirements.txt
beautifulsoup4
PyYAML
roman
Jinja2